server/requirements.txt
annotated-types==0.7.0
anthropic==0.46.0
anyio==4.8.0
appdirs==1.4.4
asgiref==3.8.1
beautifulsoup4==4.12.3
blinker==1.7.0
blis==1.2.0
boto3==1.36.26
botocore==1.36.26
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.2.0
click==8.1.7
cloudpathlib==0.20.0
colour==0.1.5
confection==0.1.5
contourpy==1.3.0
cycler==0.12.1
cymem==2.0.11
distlib==0.3.9
distro==1.9.0
dj-database-url==2.3.0
Django==5.1.6
django-cors-headers==4.7.0
django-environ==0.12.0
django-rest-authtoken==2.1.4
django-storages==1.14.5
djangorestframework==3.15.2
dotenv==0.9.9
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
fastapi==0.115.8
filelock==3.16.1
Flask==3.0.3
Flask-Cors==4.0.0
fonttools==4.54.1
frozendict==2.4.2
fsspec==2024.10.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.26.2
idna==3.4
itsdangerous==2.2.0
Jinja2==3.1.3
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
kiwisolver==1.4.7
langcodes==3.5.0
language_data==1.3.0
lxml==5.2.1
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
multitasking==0.0.11
murmurhash==1.0.12
mysql==0.0.3
mysqlclient==2.2.7
networkx==3.4.2
numpy==1.26.4
openai==1.63.2
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
peewee==3.17.3
Pillow==10.0.0
platformdirs==4.3.6
preshed==3.0.9
pydantic==2.10.6
pydantic_core==2.27.2
pygame==2.3.0
Pygments==2.19.1
PyMySQL==1.1.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
requests==2.31.0
rich==13.9.4
s3transfer==0.11.2
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
shellingham==1.5.4
six==1.16.0
skleanr==4.2.9
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.5
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.2
srsly==2.5.1
starlette==0.45.3
stripe==11.3.0
sympy==1.13.1
thinc==8.3.4
threadpoolctl==3.5.0
timedelta==2020.12.3
timm==1.0.11
tkmacosx==1.0.5
torch==2.5.1
torchvision==0.20.0
tqdm==4.67.0
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.0.5
virtualenv==20.29.1
wasabi==1.1.3
weasel==0.4.1
webencodings==0.5.1
Werkzeug==3.0.2
whitenoise==6.9.0
wrapt==1.17.2
yfinance==0.2.38
